fix: poll function update
